% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_fitted.R
\name{get_fitted}
\alias{get_fitted}
\title{Get specified entry (index) from a vector (vec)}
\usage{
get_fitted(index, vec)
}
\arguments{
\item{index}{Index of the entry to be returned}

\item{vec}{Vector from which the entry is to be returned}
}
\value{
a single element of the inputted vector
}
\description{
Get specified entry (index) from a vector (vec)
}
\examples{
get_fitted(1, c(1,2,3))
}
